Introduce the consulEdsPoller.
The consulEdsPoller is a new consolidates our requests for endpoints
from consul. Prior to the introduction of this poller each grpc
connection launched a separate goroutine to qeury consul and look for
changes. The consulEdsPoller now exposes an API allowing each connection
to register and remove itself and express what services it is interested
in receiving consul service discovery endpoint data. The subscriptions
are retained in the subscriptions map and are protected by a RWMutex.

The pulse() function handles querying consul and passing the results to
the interested parties.