Move the edsStreamHandler into its own file.
For the most part this is just moving edsStreamHandler from the eds.go
file to its own file. There are some other important changes in here
though.

First we've renamed the main loop from handle to run, and run now
properly receives the root context from eds to watch the done channel.

Secondly we implement the consulResultHandler interface with the handle
method. This allows us to receive updates from the edsConsulPoller while
also selecting on our own edsStreamHandler.Context done channel. This
way if the grpc connection is aborted or we hit an error in our loop we
unblock the consulEdsPoller from blocking on our channel update.